+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/activities/widget/WidgetSettingsViewModel.kt
@@ -59,7 +59,7 @@ class WidgetSettingsViewModel(application: Application) :
             dialogState = DialogState.create("widget-color-picker") {
                 colorPickerFactory.createColorPicker(
                     onColorPicked = ::onLabelColorSelected,
-                    onCanceled = {
+                    onDismissed = {
                         dialogState?.let(::onDialogDismissed)
                     },
                     initialColor = viewState.labelColor,