+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/usecases/GetIconColorPickerDialogUseCase.kt
@@ -21,7 +21,7 @@ constructor(
                         settings.previousIconColor = color
                         onColorSelected(icon.withTint(color))
                     },
-                    onCanceled = onDismissed,
+                    onDismissed = onDismissed,
                     initialColor = settings.previousIconColor ?: Color.BLACK,
                 )
             }