+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/utils/ColorPickerFactory.kt
@@ -1,14 +1,18 @@
 package ch.rmy.android.http_shortcuts.utils
 
+import android.app.Dialog
+import android.graphics.Color
+import android.text.InputFilter
+import android.view.LayoutInflater
 import androidx.annotation.ColorInt
-import androidx.appcompat.app.AlertDialog
-import ch.rmy.android.framework.extensions.applyIfNotNull
+import androidx.core.widget.addTextChangedListener
 import ch.rmy.android.framework.extensions.runIfNotNull
-import ch.rmy.android.framework.extensions.takeUnlessEmpty
 import ch.rmy.android.framework.utils.localization.Localizable
 import ch.rmy.android.http_shortcuts.R
-import com.skydoves.colorpickerview.ColorPickerDialog
-import com.skydoves.colorpickerview.listeners.ColorEnvelopeListener
+import ch.rmy.android.http_shortcuts.databinding.DialogColorPickerBinding
+import ch.rmy.android.http_shortcuts.utils.ColorUtil.colorIntToHexString
+import ch.rmy.android.http_shortcuts.utils.ColorUtil.hexStringToColorInt
+import com.skydoves.colorpickerview.listeners.ColorListener
 import javax.inject.Inject
 
 class ColorPickerFactory
@@ -19,35 +23,69 @@ constructor(
 
     fun createColorPicker(
         onColorPicked: (Int) -> Unit,
-        onCanceled: () -> Unit,
+        onDismissed: () -> Unit,
         title: Localizable = Localizable.EMPTY,
         @ColorInt initialColor: Int? = null,
-    ): AlertDialog {
+    ): Dialog {
         val activity = activityProvider.getActivity()
-        return ColorPickerDialog.Builder(activity)
-            .runIfNotNull(title.localize(activity).takeUnlessEmpty()) {
-                setTitle(it)
+        val binding = DialogColorPickerBinding.inflate(LayoutInflater.from(activity))
+
+        binding.colorPickerView.attachBrightnessSlider(binding.brightnessSlideBar)
+        var suppressTextWatcher = false
+        var selectedColor = initialColor ?: Color.BLACK
+        binding.colorPickerView.setInitialColor(selectedColor)
+
+        fun updateColorInputView() {
+            suppressTextWatcher = true
+            val selectionStart = binding.inputColor.selectionStart
+            val selectionEnd = binding.inputColor.selectionEnd
+            binding.inputColor.setText(selectedColor.colorIntToHexString())
+            suppressTextWatcher = false
+            if (selectionStart != -1 && selectionEnd != -1) {
+                binding.inputColor.setSelection(selectionStart, selectionEnd)
+            } else {
+                binding.inputColor.setSelection(6)
             }
-            .setPositiveButton(
-                R.string.dialog_ok,
-                ColorEnvelopeListener { envelope, fromUser ->
-                    if (fromUser) {
-                        onColorPicked(envelope.color)
-                    }
+            binding.inputColor.setBackgroundColor(selectedColor)
+        }
+
+        binding.inputColor.filters = arrayOf(
+            InputFilter { source, _, _, _, _, _ -> source.replace("[^A-Fa-f0-9]".toRegex(), "").uppercase() },
+            InputFilter.LengthFilter(6),
+        )
+        binding.inputColor.addTextChangedListener { text ->
+            if (!suppressTextWatcher) {
+                val color = text?.toString()
+                    ?.takeIf { it.length == 6 }
+                    ?.hexStringToColorInt()
+                if (color != null) {
+                    selectedColor = color
+                    binding.colorPickerView.selectByHsvColor(color)
+                    updateColorInputView()
                 }
-            )
-            .setNegativeButton(R.string.dialog_cancel) { dialogInterface, _ ->
-                dialogInterface.dismiss()
             }
-            .attachAlphaSlideBar(false)
-            .attachBrightnessSlideBar(true)
-            .setBottomSpace(12)
-            .setOnDismissListener {
-                onCanceled()
+        }
+
+        binding.colorPickerView.setColorListener(object : ColorListener {
+            override fun onColorSelected(color: Int, fromUser: Boolean) {
+                selectedColor = color
+                updateColorInputView()
             }
-            .applyIfNotNull(initialColor) {
-                colorPickerView.setInitialColor(it)
+        })
+        updateColorInputView()
+
+        return DialogBuilder(activity)
+            .runIfNotNull(title) {
+                title(it)
             }
-            .create()
+            .view(binding.root)
+            .positive(R.string.dialog_ok) {
+                onColorPicked(selectedColor)
+            }
+            .negative(R.string.dialog_cancel)
+            .dismissListener {
+                onDismissed()
+            }
+            .build()
     }
 }