+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/variables/types/ColorType.kt
@@ -37,7 +37,7 @@ class ColorType : BaseVariableType() {
                 onColorPicked = { color ->
                     emitter.onSuccess(color.colorIntToHexString())
                 },
-                onCanceled = {
+                onDismissed = {
                     if (!emitter.isDisposed) {
                         emitter.cancel()
                     }